Refactored App into index
